refactor: move OpenFeedback component to viewmodel module.
* Create a new module for ui models shared between viewmodel and m3 modules.
* Move OpenFeedback component from m3 to viewmodel modules.
* m3 module doesn't know anymore viewmodel and openfeedback modules.
* Reduce the complexity of ViewModel class.
