+++ openfeedback-m3/build.gradle.kts
@@ -12,14 +12,10 @@ library(
     kotlinMultiplatformExtension.sourceSets {
         findByName("commonMain")!!.apply {
             dependencies {
-                api(projects.openfeedback)
                 api(projects.openfeedbackResources)
-                api(projects.openfeedbackViewmodel)
-
-                api(libs.androidx.lifecycle.viewmodel.compose)
+                api(projects.openfeedbackUiModels)
 
                 implementation(kotlinMultiplatformExtension.compose.material3)
-                implementation(kotlinMultiplatformExtension.compose.ui)
             }
         }
         val androidMain by getting {
