+++ openfeedback-m3/src/androidMain/kotlin/io/openfeedback/m3/VoteItemsPreview.kt
@@ -9,8 +9,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import io.openfeedback.viewmodels.models.UIDot
-import io.openfeedback.viewmodels.models.UIVoteItem
+import io.openfeedback.ui.models.UIDot
+import io.openfeedback.ui.models.UIVoteItem
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Preview
