+++ openfeedback-m3/src/commonMain/kotlin/io/openfeedback/m3/CommentItems.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import io.openfeedback.resources.LocalStrings
-import io.openfeedback.viewmodels.models.UIComment
+import io.openfeedback.ui.models.UIComment
 
 @Composable
 internal fun CommentItems(
