+++ openfeedback-m3/src/commonMain/kotlin/io/openfeedback/m3/DotModifier.kt
@@ -6,7 +6,7 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.Fill
 import androidx.compose.ui.unit.dp
-import io.openfeedback.viewmodels.models.UIDot
+import io.openfeedback.ui.models.UIDot
 
 fun Modifier.drawDots(dots: List<UIDot>): Modifier = drawBehind {
     dots.forEach { dot ->
