+++ openfeedback-viewmodel/build.gradle.kts
@@ -13,9 +13,14 @@ library(
         getByName("commonMain") {
             dependencies {
                 implementation(projects.openfeedback)
+                api(projects.openfeedbackM3)
+                api(projects.openfeedbackUiModels)
+
+                implementation(kotlinMultiplatformExtension.compose.material3)
                 implementation(kotlinMultiplatformExtension.compose.runtime)
                 // Not sure why this is needed 🤷
                 implementation(libs.jetbrains.kotlin.stdlib)
+
                 api(libs.androidx.lifecycle.viewmodel.compose)
                 api(libs.vanniktech.multiplatform.locale)
             }
