+++ sample-app-android/src/main/java/io/openfeedback/android/MainApplication.kt
@@ -2,8 +2,8 @@ package io.openfeedback.android
 
 import android.app.Application
 import android.content.Context
-import io.openfeedback.m3.initializeOpenFeedback
 import io.openfeedback.viewmodels.OpenFeedbackFirebaseConfig
+import io.openfeedback.viewmodels.initializeOpenFeedback
 
 class MainApplication: Application() {
     lateinit var context: Context
