+++ settings.gradle.kts
@@ -13,6 +13,7 @@ includeBuild("build-logic")
 include(
     ":openfeedback",
     ":openfeedback-viewmodel",
+    ":openfeedback-ui-models",
     ":openfeedback-m3",
     ":openfeedback-resources",
     ":sample-app-android",
